Add components and route to show a modal for a phrase
In the /browse/ section only. Modal is not retreiving data
yet, just showing up and dismissing (passably). This commit
will have a few things that ought to be cleaned up and clarified
later on, e.g. the use of named functions to bind classNames to
changes in state of the component.