Put modal controls in DeckItem instead of router
Previously the app was stacking routes so that the Deck's
URL would match and the Deck would load, and then the Card's
URL would also match, rendering the modal, so the modal didn't
need to have any state.

This change is starting to get the new DeckItem component to talk
to the Modal component to manage what's open and what's closed.
It works, but it's not very good. I would rather let both parent
and child be reactive to and in control of a single `state.open`.